changed dinamic data loading way
