+++ AI Helper/Manage/ModeSwitch/ModeSwitcherBase.cs
@@ -49,14 +49,14 @@ namespace AIHelper.Manage.ModeSwitch
 
             await Task.Run(() => Action()).ConfigureAwait(true);
 
-            try
-            {
-                ManageSettings.MainForm.MOCommonModeSwitchButton.Text = MOmode ? T._("MOToCommon") : T._("CommonToMO");
-            }
-            catch (Exception ex)
-            {
-                _log.Info("An error occured while change mode switch button text: \r\n" + ex);
-            }
+            //try
+            //{
+            //    ManageSettings.MainForm.MOCommonModeSwitchButton.Text = MOmode ? T._("MOToCommon") : T._("CommonToMO");
+            //}
+            //catch (Exception ex)
+            //{
+            //    _log.Info("An error occured while change mode switch button text: \r\n" + ex);
+            //}
 
             ManageSettings.MainForm.UpdateData();
